+++ R/normalise.R
@@ -6,8 +6,8 @@
 #'
 #' @template arg_data
 #'
-#' @param to.range Normalise values to this range. If the objective is
-#'   maximised, it is normalised to \code{c(to.range[1], to.range[0])}
+#' @param to_range Normalise values to this range. If the objective is
+#'   maximised, it is normalised to \code{c(to_range[1], to_range[0])}
 #'   instead.
 #'
 #' @param lower,upper Bounds on the values. If NA, the maximum and minimum
@@ -27,10 +27,10 @@
 #'
 #' head(normalise(SPEA2minstoptimeRichmond[, 1:2], maximise = c(FALSE, TRUE)))
 #'
-#' head(normalise(SPEA2minstoptimeRichmond[, 1:2], to.range = c(0,1), maximise = c(FALSE, TRUE)))
+#' head(normalise(SPEA2minstoptimeRichmond[, 1:2], to_range = c(0,1), maximise = c(FALSE, TRUE)))
 #' 
 #' @export
-normalise <- function(data, to.range = c(1, 2), lower = NA, upper = NA, maximise = FALSE)
+normalise <- function(data, to_range = c(1, 2), lower = NA, upper = NA, maximise = FALSE)
 {
   data <- check_dataset(data)
   nobjs <- ncol(data)
@@ -45,14 +45,14 @@ normalise <- function(data, to.range = c(1, 2), lower = NA, upper = NA, maximise
   upper[no.upper] <- minmax[no.upper, 2L]
   maximise <- as.logical(rep_len(maximise, nobjs))
 
-  if (length(to.range) != 2L)
-    stop("to.range must be a vector of length 2")
+  if (length(to_range) != 2L)
+    stop("to_range must be a vector of length 2")
   
   z <- t(.Call(normalise_C,
                  as.double(t(data)),
                  as.integer(nobjs),
                  as.integer(npoints),
-                 as.double(to.range),
+                 as.double(to_range),
                  lower, upper, maximise))
   colnames(z) <- colnames(data)
   rownames(z) <- rownames(data)