+++ tests/testthat/test-normalise.R
@@ -10,7 +10,7 @@ test_that("normalise", {
   expect_equal(normalise(x, maximise = c(FALSE,TRUE)),
                my.2d.matrix(1, 1.5, 2, 2, 1.5, 1))
 
-  expect_equal(normalise(x, to.range = c(0, 1), maximise = c(FALSE,TRUE)),
+  expect_equal(normalise(x, to_range = c(0, 1), maximise = c(FALSE,TRUE)),
                my.2d.matrix(0, 0.5, 1, 1, 0.5, 0))
 
   expect_equal(normalise(my.2d.matrix(1,1,2,2)), my.2d.matrix(1,1,1,1))